requirements.txt
cloudscraper
rgbprint
pandas
scikit-learn
requests
numpy
websocket-client